requirements.txt
anyio==4.9.0
click==8.2.0
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
idna==3.10
pycryptodome==3.23.0
pydantic==1.10.13
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
teradatasql==20.0.0.30
typing_extensions==4.13.2
uvicorn==0.34.2
